ARROW-3906: [C++] Break out builder.cc into multiple compilation units
This improves readability and can also make incremental compilation faster.
